05_gcp_datastore/requirements.txt
flask
google-cloud-datastore
google-cloud-firestore
gunicorn
